style: 패키지 네이밍 컨벤션 적용
