+++ src/main/java/janggi/Application.java
@@ -2,7 +2,7 @@ package janggi;
 
 import janggi.controller.JanggiController;
 import janggi.domain.JanggiGameManager;
-import janggi.domain.Turn.ChoTurn;
+import janggi.domain.turn.ChoTurn;
 import janggi.domain.board.Board;
 import janggi.domain.strategy.BasicPlacementStrategy;
 import janggi.infrastructure.JDBCBoardRepository;
